requirements.txt
# Core Python Libraries
numpy>=1.20.0
pandas>=1.3.0
matplotlib>=3.4.0
pillow>=9.0.0

# Machine Learning & Computer Vision
ultralytics>=8.0.0  # For YOLOv8
opencv-python-headless>=4.5.0
roboflow>=1.0.0
scikit-learn>=1.0.0

# Deep Learning Libraries
torch>=1.9.0
torchvision>=0.10.0

# Data Visualization
plotly>=5.5.0
dash>=2.6.0
dash-bootstrap-components>=1.2.0
dash-core-components>=2.0.0
dash-html-components>=2.0.0
dash-table>=5.0.0

# Jupyter Environment
jupyter>=1.0.0
ipykernel>=6.0.0

# Database & API
supabase>=0.5.8
python-dotenv>=0.19.0

# Utilities
tqdm>=4.62.0
requests>=2.26.0

# Development & Testing
pytest>=7.0.0
black>=22.0.0
flake8>=4.0.0